mbe-rpi/requirements.txt
asyncio
datetime
numpy
psutil
pyserial
RPi.GPIO
smbus
websockets